+++ src/geoms/geom_bar.jl
@@ -1,116 +1,80 @@
-function handle_position(dicts)
-    handle_position(dicts[1], dicts[2])
-end
+function handle_position(aes_dict::Dict{String, Symbol}, 
+    args_dict::Dict{Any, Any}, required_aes::Vector{String}, plot_data::DataFrame)
+    # handles defaults and grouping for geom_bar/col
 
-function handle_position(aes_dict, args_dict)
-    # handles defaults and grouping for geom_bar 
+    split_var = nothing
 
     if haskey(args_dict, "position")
         if args_dict["position"] == "dodge"
             if haskey(aes_dict, "group")
                 aes_dict["dodge"] = aes_dict["group"]
+                split_var = aes_dict["dodge"]
             elseif haskey(aes_dict, "colour")
                 aes_dict["dodge"] = aes_dict["colour"]
+                split_var = aes_dict["dodge"]
             elseif haskey(aes_dict, "color")
                 aes_dict["dodge"] = aes_dict["color"]
+                split_var = aes_dict["dodge"]
             end
         elseif args_dict["position"] != "none"
             if haskey(aes_dict, "group")
                 aes_dict["stack"] = aes_dict["group"]
+                split_var = aes_dict["stack"]
             elseif haskey(aes_dict, "colour")
                 aes_dict["stack"] = aes_dict["colour"]
+                split_var = aes_dict["stack"]
             elseif haskey(aes_dict, "color")
                 aes_dict["stack"] = aes_dict["color"]
+                split_var = aes_dict["stack"]
             end
         end
     else
         if haskey(aes_dict, "group")
             aes_dict["stack"] = aes_dict["group"]
+            split_var = aes_dict["stack"]
         elseif haskey(aes_dict, "colour")
             aes_dict["stack"] = aes_dict["colour"]
+            split_var = aes_dict["stack"]
         elseif haskey(aes_dict, "color")
             aes_dict["stack"] = aes_dict["color"]
+            split_var = aes_dict["stack"]
         end
     end
 
-    return (aes_dict, args_dict)
+    # for geom_bar, we need to summarize counts
+    if args_dict["geom_name"] == "geom_bar"
+        if haskey(aes_dict, "x") && !haskey(aes_dict, "y")
+            grouping_var = Symbol(aes_dict["x"])
+            aes_dict["y"] = :count
+            required_aes = ["x", "y"]
+        elseif haskey(aes_dict, "y") && !haskey(aes_dict, "x")
+            grouping_var = Symbol(aes_dict["y"])
+            aes_dict["x"] = :count
+            args_dict["direction"] = "x"
+            required_aes = ["y", "x"]
+        else
+            @error "geom_bar requires either an x or y aesthetic, but not both."
+        end
+
+        if !isnothing(split_var)
+            plot_data = @chain plot_data begin
+                groupby([grouping_var, split_var])
+                @summarize(count = n())
+                @ungroup
+            end
+        else
+            plot_data = @chain plot_data begin
+                sort(grouping_var)
+                groupby(grouping_var)
+                @summarize(count = n())
+                @ungroup
+            end
+        end
+    end
+
+    return (aes_dict, args_dict, required_aes, plot_data)
 end
 
-
-"""
-    geom_bar(aes(...), ...)
-    geom_bar(plot::GGPlot, aes(...), ...)
-    
-    Represent data as bars. 
-
-    # Arguments
-
-    - plot::GGPlot (optional): a plot object to "add" this geom to
-    - `aes(...)`: the names of the columns in the plot DataFrame that will be used to decide where the points are plotted.
-    - `...`: options that are not mapped to a column 
-
-    # Required Aesthetics
-
-    - x
-    - y
-
-    # Supported Optional Aesthetics
-
-    - alpha
-    - colour/color
-    - fill
-    - group
-    - linetype
-    - linewidth
-
-    # Supported Options
-
-    - alpha
-    - colour/color
-    - fill
-    - group
-    - linetype
-    - linewidth
-
-"""
-geom_bar = geom_template("geom_bar", ["x"], Makie.BarPlot,  AlgebraOfGraphics.frequency(); dict_function = handle_position)
-
-"""
-    geom_histogram(aes(...), ...)
-    geom_histogram(plot::GGPlot, aes(...), ...)
-    
-    Represent data as a histogram. 
-
-    # Arguments
-
-    - plot::GGPlot (optional): a plot object to "add" this geom to
-    - `aes(...)`: the names of the columns in the plot DataFrame that will be used to decide where the points are plotted.
-    - `...`: options that are not mapped to a column 
-
-    # Required Aesthetics
-
-    - x
-
-    # Supported Optional Aesthetics
-
-    - alpha
-    - colour/color
-    - fill
-    - group
-    - linetype
-    - linewidth
-
-    # Supported Options
-
-    - alpha
-    - colour/color
-    - fill
-    - group
-    - linetype
-    - linewidth
-
-"""
-geom_histogram = geom_template("geom_histogram", ["x"], nothing, AlgebraOfGraphics.histogram())
 
 """
     geom_col(aes(...), ...)
@@ -122,7 +86,8 @@ geom_histogram = geom_template("geom_histogram", ["x"], nothing, AlgebraOfGraphi
 
     - plot::GGPlot (optional): a plot object to "add" this geom to
     - `aes(...)`: the names of the columns in the plot DataFrame that will be used to decide where the points are plotted.
-    - `...`: options that are not mapped to a column 
+    - `position::String`: "stack" (the default) or "dodge"
+    - `...`: options that are not mapped to a column (passed to Makie.BarPlot) 
 
     # Required Aesthetics
 
@@ -148,4 +113,43 @@ geom_histogram = geom_template("geom_histogram", ["x"], nothing, AlgebraOfGraphi
     - linewidth
 
 """
-geom_col = geom_template("geom_col", ["x", "y"], Makie.BarPlot, mapping(); dict_function = handle_position)
+geom_col = geom_template("geom_col", ["x", "y"], :BarPlot; aes_function = handle_position)
+"""
+    geom_col(aes(...), ...)
+    geom_col(plot::GGPlot, aes(...), ...)
+    
+    Represent data as bars. 
+
+    # Arguments
+
+    - plot::GGPlot (optional): a plot object to "add" this geom to
+    - `aes(...)`: the names of the columns in the plot DataFrame that will be used to decide where the points are plotted.
+    - `...`: options that are not mapped to a column (passed to Makie.BarPlot)
+
+    # Required Aesthetics
+
+    - x OR y
+
+    # Supported Optional Aesthetics (See aes() for specification options)
+
+    - alpha
+    - colour/color
+    - fill
+    - group
+    - linetype
+    - linewidth
+
+    # Supported Options
+
+    - alpha::Float32
+    - position::String - "stack" (the default) or "dodge"
+    - colour/color
+    - fill
+    - group
+    - linetype
+    - linewidth
+
+"""
+geom_bar = geom_template("geom_bar", String[], :BarPlot; aes_function = handle_position)
+
+geom_histogram = geom_template("geom_histogram", ["x"], :Hist)