+++ src/geoms/geom_contour.jl
@@ -1,2 +1,2 @@
-geom_tile = geom_template("geom_tile", ["x", "y"], Makie.Heatmap, AlgebraOfGraphics.density())
-geom_contour = geom_template("geom_contour", ["x", "y"], Makie.Contour, AlgebraOfGraphics.density())
+geom_tile = geom_template("geom_tile", ["x", "y", "z"], :Heatmap)
+geom_contour = geom_template("geom_contour", ["x", "y", "z"], :Contour)