+++ src/geoms/geom_errorbar.jl
@@ -1,29 +1,3 @@
-function geom_errorbar(plot::GGPlot, args...; kwargs...)
-    aes_dict, args_dict = extract_aes(args, kwargs)
-
-    args_dict["geom_name"] = "geom_errorbar"
-    
-    return plot + build_geom(aes_dict, args_dict, 
-        ["x", "ymin", "ymax"], # required aesthetics
-        Makie.Rangebars, # function for visual layer
-        AlgebraOfGraphics.mapping(), # function for analysis layer
-        special_aes = Dict("width" => "whiskerwidth")) 
-end
-
-function geom_errorbarh(plot::GGPlot, args...; kwargs...)
-    aes_dict, args_dict = extract_aes(args, kwargs)
-
-    args_dict["geom_name"] = "geom_errorbarh"
-    args_dict["errorbar_direction"] = :x
-
-    return plot + build_geom(aes_dict, args_dict, 
-        ["y", "xmin", "xmax"], # required aesthetics
-        Makie.Rangebars, # function for visual layer
-        AlgebraOfGraphics.mapping(), # function for analysis layer
-        special_aes = Dict("width" => "whiskerwidth")) 
-
-end
-
 function geom_errorbar(args...; kwargs...)
     aes_dict, args_dict = extract_aes(args, kwargs)
 
@@ -31,8 +5,8 @@ function geom_errorbar(args...; kwargs...)
     
     return build_geom(aes_dict, args_dict, 
         ["x", "ymin", "ymax"], # required aesthetics
-        Makie.Rangebars, # function for visual layer
-        AlgebraOfGraphics.mapping(), # function for analysis layer
+        :Rangebars, # function for visual layer
+        do_nothing;
         special_aes = Dict("width" => "whiskerwidth")) 
 end
 
@@ -44,8 +18,16 @@ function geom_errorbarh(args...; kwargs...)
 
     return build_geom(aes_dict, args_dict, 
         ["y", "xmin", "xmax"], # required aesthetics
-        Makie.Rangebars, # function for visual layer
-        AlgebraOfGraphics.mapping(), # function for analysis layer
+        :Rangebars, # function for visual layer
+        do_nothing;
         special_aes = Dict("width" => "whiskerwidth")) 
 
+end
+
+function geom_errorbarh(plot::GGPlot, args...; kwargs...)
+    return plot + geom_errorbarh(args..., kwargs...)
+end
+
+function geom_errorbar(plot::GGPlot, args...; kwargs...)
+    return plot + geom_errorbar(args..., kwargs...)
 end