+++ src/geoms/geom_text.jl
@@ -1,7 +1,2 @@
-geom_text = geom_template("geom_text", ["x", "y"], Makie.Text, AlgebraOfGraphics.mapping(); 
-    extra_args = Dict(["text_function" => AlgebraOfGraphics.verbatim, 
-                      "label_function" => AlgebraOfGraphics.verbatim]))
-
-geom_text = geom_template("geom_text", ["x", "y"], Makie.Text, AlgebraOfGraphics.mapping(); 
-    extra_args = Dict(["text_function" => AlgebraOfGraphics.verbatim, 
-                      "label_function" => AlgebraOfGraphics.verbatim]))
+geom_text = geom_template("geom_text", ["x", "y"], :Text)
+geom_label = geom_template("geom_text", ["x", "y"], :Text)