+++ src/ggplot.jl
@@ -1,18 +1,17 @@
 function ggplot(args...; kwargs...)
     aes_dict, args_dict = extract_aes(args, kwargs)
 
-    if !haskey(args_dict, "height")
-        args_dict["height"] = 400
-    end
-    
-    if !haskey(args_dict, "width")
-        args_dict["width"] = 600
-    end
-
     if haskey(args_dict, "data")
-        plot_data = AlgebraOfGraphics.data(args_dict["data"])
+        if args_dict["data"] isa DataFrame
+            plot_data = args_dict["data"]
+            delete!(args_dict, "data")
+        else
+            type = typeof(args_dict["data"])
+            @warn "Data was provided in ggplot function with unsupported type: $type. Data argument ignored."
+            plot_data = nothing
+        end
     else
-        plot_data = mapping()
+        plot_data = nothing
     end
 
     GGPlot([], 
@@ -25,19 +24,9 @@ end
 function ggplot(data::DataFrame, args...; kwargs...)
     aes_dict, args_dict = extract_aes(args, kwargs)
 
-    if !haskey(args_dict, "height")
-        args_dict["height"] = 400
-    end
-    
-    if !haskey(args_dict, "width")
-        args_dict["width"] = 600
-    end
-
-    plot_data = AlgebraOfGraphics.data(data)
-
     GGPlot([], 
            aes_dict, 
-           plot_data,
+           data,
            args_dict, 
            theme_ggplot2()) 
 end