+++ src/ggsave.jl
@@ -1,4 +1,4 @@
-function ggsave(path::String, plot::GGPlot; scale = 0.75)
+function ggsave(path::String, plot::GGPlot; scale = 2)
     save(path, 
     with_theme(plot.theme) do
         draw_ggplot(plot)
@@ -6,7 +6,7 @@ function ggsave(path::String, plot::GGPlot; scale = 0.75)
     px_per_unit = scale)
 end
 
-function ggsave(plot::GGPlot, path::String; scale = 0.75)
+function ggsave(plot::GGPlot, path::String; scale = 2)
     save(path, 
         with_theme(plot.theme) do
             draw_ggplot(plot)