+++ src/scales.jl
@@ -14,7 +14,7 @@ function continuous_scale_to_ggoptions(args_dict::Dict)
         options_dict[args_dict["scale"] * "reversed"] = args_dict["reversed"]
     end
 
-    return(options_dict)
+    return AxisOptions(Dict(Symbol(k) => v for (k, v) in options_dict))
 
 end
 
@@ -44,8 +44,7 @@ function colour_scale_to_ggoptions(args_dict::Dict)
         end
     end
 
-    return(options_dict)
-
+    return AxisOptions(Dict(Symbol(k) => v for (k, v) in options_dict))
 end
 
 # Generator - generates two function signatures
@@ -93,16 +92,16 @@ scale_color_continuous = scale_template("colour_continuous", colour_scale_to_ggo
 
 scale_x_continuous = scale_template("x", continuous_scale_to_ggoptions)
 scale_y_continuous = scale_template("y", continuous_scale_to_ggoptions)
-scale_x_log10 = scale_template("x", continuous_scale_to_ggoptions; trans = AlgebraOfGraphics.log10)
-scale_y_log10 = scale_template("y", continuous_scale_to_ggoptions; trans = AlgebraOfGraphics.log10)
+scale_x_log10 = scale_template("x", continuous_scale_to_ggoptions; trans = Makie.log10)
+scale_y_log10 = scale_template("y", continuous_scale_to_ggoptions; trans = Makie.log10)
 scale_x_reverse = scale_template("x", continuous_scale_to_ggoptions; reverse = true)
 scale_y_reverse = scale_template("y", continuous_scale_to_ggoptions; reverse = true)
-scale_x_sqrt = scale_template("x", continuous_scale_to_ggoptions; trans = AlgebraOfGraphics.sqrt)
-scale_y_sqrt = scale_template("y", continuous_scale_to_ggoptions; trans = AlgebraOfGraphics.sqrt)
-scale_x_log2 = scale_template("x", continuous_scale_to_ggoptions; trans = AlgebraOfGraphics.log2)
-scale_y_log2 = scale_template("y", continuous_scale_to_ggoptions; trans = AlgebraOfGraphics.log2)
-scale_x_log = scale_template("x", continuous_scale_to_ggoptions; trans = AlgebraOfGraphics.log)
-scale_y_log = scale_template("y", continuous_scale_to_ggoptions; trans = AlgebraOfGraphics.log)
+scale_x_sqrt = scale_template("x", continuous_scale_to_ggoptions; trans = Makie.sqrt)
+scale_y_sqrt = scale_template("y", continuous_scale_to_ggoptions; trans = Makie.sqrt)
+scale_x_log2 = scale_template("x", continuous_scale_to_ggoptions; trans = Makie.log2)
+scale_y_log2 = scale_template("y", continuous_scale_to_ggoptions; trans = Makie.log2)
+scale_x_log = scale_template("x", continuous_scale_to_ggoptions; trans = Makie.log)
+scale_y_log = scale_template("y", continuous_scale_to_ggoptions; trans = Makie.log)
 scale_x_logit = scale_template("x", continuous_scale_to_ggoptions; trans = Makie.logit)
 scale_y_logit = scale_template("y", continuous_scale_to_ggoptions; trans = Makie.logit)
 scale_x_pseudolog10 = scale_template("x", continuous_scale_to_ggoptions; trans = Makie.pseudolog10)