+++ src/structs.jl
@@ -1,22 +1,27 @@
 struct Geom
     aes::Dict
     args::Dict
-    data::Layer
-    visual::Layer
-    analysis::Layer
     required_aes::AbstractArray
-    optional_aes::Dict
+    special_aes::Dict
+    data::Union{Nothing, DataFrame}
+    visual::Symbol
     axis_options::Dict
+    aes_function::Function
 end
 
 struct GGPlot
     geoms::AbstractArray
     default_aes::Dict
-    data::Layer
+    data::Union{Nothing, DataFrame}
     axis_options::Dict
     theme::Attributes
 end
 
 struct Aesthetics
-    aes::Dict
+    positional::AbstractArray
+    named::Dict
+end
+
+struct AxisOptions
+    opt::Dict{Symbol, Any}
 end